requirements.txt
firecrawl-py
unstructured[local-inference]
openai
pymongo
boto3
python-dotenv